Deleting last 2 rows should feel like "clearing" the rows
Previously:
Given only 2 rows in the list
When the first row is deleted
Then the second row moves up to the first row,
and a new row is added to the end to the second row position.

Now:
Given only 2 rows in the list
When the first row is deleted
Then a new row is added back into the first row,
and the second row is not affected.

This simplifies the user's impression of the animation.