Polish split bills (#155)
* Add remove user handler in split bill page

* Add delete row button to split bill page

* Implement drag-n-drop reordering of split bill people

* Fix drag helper list style and add cursor for drag handle

* Style Split-Bill page according to new design mockups

- Open Sans added

* Allow placeholder bill title to be submitted

* Animate split-bill page when list changes

* Remove toggling behaviour for split bill toggle buttons

* Mark person's name required, and bill title not required

* Always show at least 2 rows in split bill form

* Implement split-bill inline error validation

- Introduction of ErrorBubble components
- Introduction of the errors object
- Create a "fake" progress feedback delay

* Don't reset payee when selected payer button is re-selected

The "Payer" button should only do something if that person is a "payee".
If the person is already a payer, pressing the "payer" button should not
clear someone else's payee status.

* Fix error bubble font weight

Bubble should use normal italics, but only light italics were downloaded
and the browser was forced to use artificial boldening.

* Fix bug where deleting a payee row keeps the payee

* Deleting last 2 rows should feel like "clearing" the rows

Previously:
Given only 2 rows in the list
When the first row is deleted
Then the second row moves up to the first row,
and a new row is added to the end to the second row position.

Now:
Given only 2 rows in the list
When the first row is deleted
Then a new row is added back into the first row,
and the second row is not affected.

This simplifies the user's impression of the animation.

* Refactor split-bill file structure to best practices

I.e. dedicated folder for Split, and ErrorBubble

* Prevent error bubble in row being overlapped by row below

This bug happened in inconsistent cases.
See comment in this commit.

* Trim whitespace off split bill textfields

* Fix fluid input not resizing with white spaces in input

* Gracefully handle large inputs in split bill form

A very long title should not push the total text field outside of the
box.

* Fix Safari focus stuck on previous textfield bug

* Flatten FluidInput's fake span structure

This is in attempt to simplify FluidInput's layout behaviour, hopefully
making it easier to debug CSS problems due to misalignments and
overflows.

* Allow header error to overflow outside, and fix FluidInput alignment

The cause of alignment problems was the padding we used to fix the
clipped descenders in the text fields.

The clipping of the descenders are now fixed by resetting the height and
line-height attributes.

* Reduce flip animation distortion due to border thickness

Distortion was noticeable on Safari, and in some cases, other browsers
too. To reproduce the distortion, create a list of people, then try
deleting a row. As you delete a row, you may notice that the header text
moves slightly even when it should remain still.